Changed the wall thickness to 2 inches (standard) and height to 42 inches (7 bricks tall) and left the length at 8 feet (243-cm).
Changed out the beamline componenets to match the latest update on GitHub from the G4 SBS code.  This helped a TON as it includes the proper dipole magnet models for the corrector magnets, magnetic flux shielding around the beamline, and the size of the beamline.

It correctly goes back to the target chamber as well.